azyroth/project_template/requirements.txt
alembic==1.16.4
blinker==1.9.0
click==8.1.8
flask==3.1.1
greenlet==3.2.3
gunicorn==23.0.0
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
PyMySQL==1.1.1
python-dotenv==1.1.1
sqlalchemy==2.0.41
tomli==2.2.1
typing-extensions==4.14.1
werkzeug==3.1.3
zipp==3.23.0
